Give isf-panel service a low oom_adj value to avoid killed
Currently, a positive oom_adj make isf-panel service likely to be killed
by system-server, which will pick up the one with positive oom_adj and
increase it's oom_adj whenever it go to inactive or backgrd. Once
its oom_adj larger or equal 15, it got killed by system-server.

For sysvinit, there is a kernel module named pmon and user space
tool which set important services as vip or permenant processes. So
they're lesser to be killed and will be relaunched once it got killed.

In systemd environment, it's easy to implement this feature by
leverage the systemd feature.

Change-Id: Ic4e99112b5f7328c967e8ce987b6fc367745493a
Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>