efl_immodule: Make sure we call ecore_x_init/shutdown when we use ecore_x functions.
This fixes P130329-9887. If we don't call ecore_x_init our usage of it
is not recorded in the refcount and it might drop to 0 and free the
_ecore_x_disp resource and crash on the next ecore_x function call:

(gdb) bt
0  0xb6de9860 in XGetInputFocus () from /usr/lib/libX11.so.6
1  0xb66222b0 in ecore_x_window_focus_get () at ecore_x_window.c:612
2  0xb4a85328 in check_focus_out_by_lockscreen ()
from /usr/lib/ecore/immodules/libisf-imf-module.so
3  0xb4a85328 in check_focus_out_by_lockscreen ()
from /usr/lib/ecore/immodules/libisf-imf-module.so
Backtrace stopped: previous frame identical to this frame (corrupt stack?)

Handle ecore_x_init/shutdown in the im module to ensure calls to ecore_x are safe.

Change-Id: I65d17e4d820a20c8af565549bff00654c549daaf