Do not run scim in daemon mode
In systemd environment, run service in no-daemonized mode is recommended
and that's much simple to track the service launched by systemd.

Meanwhile, this patch also fix scim.service fails to be activated due to
it fail to connect PanelAgent server currently and so timed out. And
then the whole user session timed out.

Change-Id: I840e39e9236c35c859cc972899440209b3abc500
Signed-off-by: Chengwei Yang <chengwei.yang@intel.com>