Using correct cast, specially for 64 bits
The value returned by sizeof is a count of char
(the chars having CHAR_BIT bits -almost always 8-).

Thus, the correct way of computing the offset of a pointer
in memory using the operator sizeof is to cast to char*.

The cast to unsigned int is wrong on most 64 bits architectures
because integers are often 32 bits wide.

Change-Id: I58d4673c9076fedf15378321dd8c0c8d8e6259e6
Signed-off-by: José Bollo <jose.bollo@open.eurogiciel.org>