Fixed the bug that wrong modifiers value in SCIM key event is sent to IME and Engine.
This problem doesn't happen in the simple entry app.
But when the input method is deactivated and activated continuously by moving the focus among the many entrys,
the uninitialized modifiers variable has a garbage value. Because of this IME couldn't handle the key event properly.

And the modifiers value which is sent by IME or returned by filter_function is not need to be saved in wsc_ctx->modifiers variable.
H/W key event's modifiers value should not be affected by IME's logic.

Change-Id: Id856d5ed4e876c857e879f42412df78527159426